Create inventory
The Inventory Management system is used to manage store products and keep track of all goods stock and also we can use it to check the purchases history of the store. Basically, we are developing a system to manage/automate some processes of any retail store by using the computer. So by using this system we can keep record data of products and even all purchases happened in-store.

So here we are going to develop an inventory management system using python and for storing data we are going to use JSON files. First of all, we will create sample data by using python code itself and then developing all functions one by one.

We have to open JSON files on each function and loading JSON data in python dictionary format and again reverse the process of converting dictionary data back to JSON format then closing the file.